refactor(shortint): move parameters set to their own directory
This is done to ease automatic parameters updates.
